+++ app/controllers/scores_controller.rb
@@ -24,8 +24,11 @@ class ScoresController < ApplicationController
     score = Score.find_by(name: params[:name]) ||
       raise(ActiveRecord::RecordNotFound.new("no such score name: '#{params[:name]}'"))
 
+    elements_datatable = Datatable.new(score.elements, [:number, :name, :element_type, :info, :base_value, :credit, :goe, :judges, :value])
+    components_datatable = Datatable.new(score.components, [:number, :name, :factor, :judges, :value])
+    
     respond_to do |format|
-      format.html { render locals: {score: score}}
+      format.html { render locals: {score: score, elements_datatable: elements_datatable, components_datatable: components_datatable}}
       #format.json { render :show, handlers: :jbuilder, locals: {score: score } }
       format.json { render json: score.as_json.merge({elememnts: score.elements, components: score.components })}
     end