+++ config/initializers/draper.rb
@@ -1,2 +1,3 @@
 #Draper::CollectionDecorator.delegate :current_page, :total_pages, :limit_value, :total_count
 Draper::CollectionDecorator.delegate :current_page, :total_pages, :limit_value, :total_count, :entry_name, :offset_value, :column_names
+